Windows: Experimental: Allow containerd for runtime
This is the first step in refactoring moby (dockerd) to use containerd on Windows.
Similar to the current model in Linux, this adds the option to enable it for runtime.
It does not switch the graphdriver to containerd snapshotters.

The first bit of this change refactors the libcontainerd package to a series of
subpackages so that either a "local" containerd (1) or a "remote" (2) containerd
can be loaded as opposed to conditional compile as "local" for Windows and
"remote" for Linux.

The second bit of this change updates libcontainerd such that Windows has an
option to allow the use of a "remote" containerd. Here, it communicates over
a named pipe using GRPC. This is currently guarded behind the experimental flag,
an environment variable, build number, and the providing of a pipename to
connect to containerd.

The third bit of this change are infrastructure pieces such as under pkg/system
to have helper functions for determining whether containerd is being used.

(1) "local" containerd is what the daemon on Windows has used since inception.
It's not really containerd at all - it's simply local invocation of HCS APIs
directly in-process from the daemon through the Microsoft/hcsshim library.

(2) "remote" containerd is what docker on Linux uses for it's runtime. It means
that there is a separate containerd service running, and docker communicates over
GRPC to it.

To try this out, you will need RS5 and to start with something like the following:

Window 1:
	containerd --log-level debug

Window 2:
	$env:DOCKER_WINDOWS_CONTAINERD=1
	dockerd --experimental -D --containerd \\.\pipe\containerd-containerd

You will need the following binaries from github.com/containerd/containerd in your path:
 - containerd.exe
 - containerd-shim-runhcs-v1.exe

You will need the following binary from github.com/Microsoft/hcsshim in your path:
 -  runhcs.exe

For LCOW, it will require and initrd.img and kernel in `C:\Program Files\Linux Containers`.
This is no different to the current requirements.

Major steps:

[x] WCOW process isolation
[x] WCOW Hyper-V isolation
[ ] LCOW (RS5+ using HCS v2 APIs, not the existing HCS v1 APIs)
[ ] Confirm that CI runs and passes locally. (It will be a while before that can be in public CI)
[ ] Use the containerd OCI packages to build the OCI spec
[ ] Extend to use containerd snapshotters rather than graphdrivers

Note that containerd and hcsshim for HCS v2 APIs do not yet support all the required
functionality needed for docker. This will come in time - this is a baby (although large)
step to migrating Docker on Windows to containerd.

Signed-off-by: John Howard <jhoward@microsoft.com>