Windows: Allow UNC bind mounts
Signed-off-by: John Howard <jhoward@microsoft.com>

Avoids needing a workaround where a symlink is created to an SMB
share on the host such as `mklink /D c:\containershare \\server\share`
and using that as the source/host-path of the bind-mount.

To use UNC paths, a global SMB mapping must be pre-created on the host.

eg
 - $cred=Get-Credential
 - New-SMBGlobalMapping -remotepath \\server\share -Credential $cred

Then docker run -v \\server\share:c:\in-container ...